+++ Cargo.toml
@@ -146,6 +146,7 @@ members = [
     "version",
     "vortexor",
     "vote",
+    "votor",
     "watchtower",
     "wen-restart",
     "zk-keygen",
@@ -565,6 +566,7 @@ solana-version = { path = "version", version = "=2.3.0" }
 solana-vote = { path = "vote", version = "=2.3.0" }
 solana-vote-interface = "=2.2.2"
 solana-vote-program = { path = "programs/vote", version = "=2.3.0", default-features = false }
+solana-votor = { path = "votor", version = "=2.3.0" }
 solana-wen-restart = { path = "wen-restart", version = "=2.3.0" }
 solana-zk-elgamal-proof-program = { path = "programs/zk-elgamal-proof", version = "=2.3.0" }
 solana-zk-keygen = { path = "zk-keygen", version = "=2.3.0" }
