+++ core/src/replay_stage.rs
@@ -3,10 +3,7 @@ use {
     crate::{
         alpenglow_consensus::{
             block_creation_loop::{LeaderWindowNotifier, ReplayHighestFrozen},
-            vote_history::VoteHistory,
-            vote_history_storage::VoteHistoryStorage,
             voting_loop::{GenerateVoteTxResult, VotingLoop, VotingLoopConfig},
-            CertificateId,
         },
         banking_stage::update_bank_forks_and_poh_recorder_for_new_tpu_bank,
         banking_trace::BankingTracer,
@@ -90,6 +87,9 @@ use {
     },
     solana_timings::ExecuteTimings,
     solana_vote::vote_transaction::VoteTransaction,
+    solana_votor::{
+        vote_history::VoteHistory, vote_history_storage::VoteHistoryStorage, CertificateId,
+    },
     std::{
         collections::{HashMap, HashSet},
         num::NonZeroUsize,
@@ -4667,9 +4667,6 @@ pub(crate) mod tests {
     use {
         super::*,
         crate::{
-            alpenglow_consensus::vote_history_storage::{
-                NullVoteHistoryStorage, VoteHistoryStorage,
-            },
             consensus::{
                 progress_map::{ValidatorStakeInfo, RETRANSMIT_BASE_DELAY_MS},
                 tower_storage::{FileTowerStorage, NullTowerStorage},
@@ -4720,6 +4717,7 @@ pub(crate) mod tests {
         solana_transaction_status::VersionedTransactionWithStatusMeta,
         solana_vote::vote_transaction,
         solana_vote_program::vote_state::{self, TowerSync, VoteStateVersions},
+        solana_votor::vote_history_storage::{NullVoteHistoryStorage, VoteHistoryStorage},
         std::{
             fs::remove_dir_all,
             iter,
