+++ core/src/validator.rs
@@ -5,12 +5,8 @@ use {
     crate::{
         accounts_hash_verifier::AccountsHashVerifier,
         admin_rpc_post_init::AdminRpcRequestMetadataPostInit,
-        alpenglow_consensus::{
-            block_creation_loop::{
-                self, BlockCreationLoopConfig, LeaderWindowNotifier, ReplayHighestFrozen,
-            },
-            vote_history::{VoteHistory, VoteHistoryError},
-            vote_history_storage::{NullVoteHistoryStorage, VoteHistoryStorage},
+        alpenglow_consensus::block_creation_loop::{
+            self, BlockCreationLoopConfig, LeaderWindowNotifier, ReplayHighestFrozen,
         },
         banking_trace::{self, BankingTracer, TraceError},
         cluster_info_vote_listener::VoteTracker,
@@ -138,6 +134,10 @@ use {
     solana_turbine::{self, broadcast_stage::BroadcastStageType},
     solana_unified_scheduler_pool::DefaultSchedulerPool,
     solana_vote_program::vote_state,
+    solana_votor::{
+        vote_history::{VoteHistory, VoteHistoryError},
+        vote_history_storage::{NullVoteHistoryStorage, VoteHistoryStorage},
+    },
     solana_wen_restart::wen_restart::{wait_for_wen_restart, WenRestartConfig},
     std::{
         borrow::Cow,
@@ -2051,7 +2051,7 @@ fn post_process_restored_tower(
 }
 
 fn post_process_restored_vote_history(
-    restored_vote_history: crate::alpenglow_consensus::vote_history_storage::Result<VoteHistory>,
+    restored_vote_history: solana_votor::vote_history_storage::Result<VoteHistory>,
     validator_identity: &Pubkey,
     config: &ValidatorConfig,
     bank_forks: &BankForks,
