+++ core/src/voting_service.rs
@@ -1,6 +1,5 @@
 use {
     crate::{
-        alpenglow_consensus::vote_history_storage::{SavedVoteHistoryVersions, VoteHistoryStorage},
         consensus::tower_storage::{SavedTowerVersions, TowerStorage},
         next_leader::upcoming_leader_tpu_vote_sockets,
         staked_validators_cache::StakedValidatorsCache,
@@ -19,6 +18,7 @@ use {
         transaction::Transaction,
         transport::TransportError,
     },
+    solana_votor::vote_history_storage::{SavedVoteHistoryVersions, VoteHistoryStorage},
     std::{
         net::SocketAddr,
         sync::{Arc, RwLock},
@@ -340,10 +340,7 @@ impl VotingService {
 mod tests {
     use {
         super::*,
-        crate::{
-            alpenglow_consensus::vote_history_storage::{NullVoteHistoryStorage, SavedVoteHistory},
-            consensus::tower_storage::NullTowerStorage,
-        },
+        crate::consensus::tower_storage::NullTowerStorage,
         alpenglow_vote::{
             bls_message::{BLSMessage, VoteMessage},
             vote::Vote,
@@ -372,6 +369,7 @@ mod tests {
             recvmmsg::recv_mmsg,
             socket::SocketAddrSpace,
         },
+        solana_votor::vote_history_storage::{NullVoteHistoryStorage, SavedVoteHistory},
         std::{
             net::SocketAddr,
             sync::{atomic::AtomicBool, Arc, RwLock},
