+++ local-cluster/Cargo.toml
@@ -44,6 +44,7 @@ solana-transaction = { workspace = true, features = ["bincode"] }
 solana-turbine = { workspace = true }
 solana-vote = { workspace = true }
 solana-vote-program = { workspace = true }
+solana-votor = { workspace = true }
 static_assertions = { workspace = true }
 strum = { workspace = true, features = ["derive"] }
 tempfile = { workspace = true }
