+++ local-cluster/src/local_cluster.rs
@@ -10,7 +10,6 @@ use {
     solana_accounts_db::utils::create_accounts_run_and_snapshot_dirs,
     solana_client::connection_cache::ConnectionCache,
     solana_core::{
-        alpenglow_consensus::vote_history_storage::FileVoteHistoryStorage,
         consensus::tower_storage::FileTowerStorage,
         validator::{Validator, ValidatorConfig, ValidatorStartProgress, ValidatorTpuConfig},
     },
@@ -59,6 +58,7 @@ use {
         vote_instruction,
         vote_state::{self, VoteInit},
     },
+    solana_votor::vote_history_storage::FileVoteHistoryStorage,
     std::{
         collections::HashMap,
         io::{Error, ErrorKind, Result},
