+++ votor/src/vote_history.rs
@@ -1,8 +1,11 @@
+#[cfg(feature = "frozen-abi")]
+use solana_frozen_abi_macro::{frozen_abi, AbiExample};
 use {
     super::vote_history_storage::{
         Result, SavedVoteHistory, SavedVoteHistoryVersions, VoteHistoryStorage,
     },
     alpenglow_vote::vote::Vote,
+    serde::{Deserialize, Serialize},
     solana_sdk::{clock::Slot, hash::Hash, pubkey::Pubkey, signature::Keypair},
     std::collections::{HashMap, HashSet},
     thiserror::Error,
@@ -48,7 +51,7 @@ impl VoteHistoryVersions {
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Default)]
 pub struct VoteHistory {
     /// The validator identity that cast votes
-    pub(crate) node_pubkey: Pubkey,
+    pub node_pubkey: Pubkey,
 
     /// The slots which this node has cast either a notarization or skip vote
     voted: HashSet<Slot>,
